feat: move AST base class to ast package
